requirements.txt
dnspython
pycryptodome
ecdsa
IPy
ipwhois
termcolor
ipaddress
connexion == 2.6.0
swagger-ui-bundle
python_dateutil == 2.6.0
setuptools >= 21.0.0
redis
